fix: Correct bounce and info window logic
Fixed bounce() to make sure that clicking on a bouncing marker bounces the marker after stopping inprogress animations. Also enhanced generateInfoWindow to allow for delay between API calls to wikipedia and opening the infoWindow. Also reformatted the infoWindow content to provide image attribution and links to the full size image.

Resolves: #3, #13